+++ cantera18/branches/scons/Cantera/src/transport/LiquidTransport.h
@@ -729,7 +729,7 @@ namespace Cantera {
     /*!
      * Either 1, 2, or 3
      */
-    int m_nDim;
+    size_t m_nDim;
 
   private:
