+++ cantera18/branches/scons/Cantera/src/zeroD/FlowDevice.cpp
@@ -22,7 +22,7 @@ namespace CanteraZeroD {
         m_nspin = mixin->nSpecies();
         m_nspout = mixout->nSpecies();
         string nm;
-        int ki, ko;
+        size_t ki, ko;
         for (ki = 0; ki < m_nspin; ki++) {
             nm = mixin->speciesName(ki);
             ko = mixout->speciesIndex(nm);
@@ -45,9 +45,9 @@ namespace CanteraZeroD {
      * Mass flow rate of outlet species k.  Returns zero if this
      * species is not present in the upstream mixture.
      */
-    doublereal FlowDevice::outletSpeciesMassFlowRate(int k) {
+    doublereal FlowDevice::outletSpeciesMassFlowRate(size_t k) {
         if (k < 0 || k >= m_nspout) return 0.0;
-        int ki = m_out2in[k];
+        size_t ki = m_out2in[k];
         if (ki < 0) return 0.0;
         return m_mdot * m_in->massFraction(ki);
     }
