+++ cantera18/branches/scons/apps/bvp/blasius.cpp
@@ -67,7 +67,7 @@ public:
     // Specify the residual. This is where the ODE system and boundary 
     // conditions are specified. The solver will attempt to find a solution
     // x so that this function returns 0 for all n and j.
-    virtual doublereal residual(doublereal* x, int n, int j) {
+    virtual doublereal residual(doublereal* x, size_t n, size_t j) {
 
         // if n = 0, return the residual for the first ODE
         if (n == 0) {
